Move mobile app to main.js
